Refactor: move custom selector models
