+++ app/src/main/java/fr/free/nrw/commons/db/AppDatabase.kt
@@ -5,7 +5,7 @@ import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
 import fr.free.nrw.commons.contributions.Contribution
 import fr.free.nrw.commons.contributions.ContributionDao
-import fr.free.nrw.commons.customselector.database.UploadedStatus
+import fr.free.nrw.commons.data.models.model.UploadedStatus
 import fr.free.nrw.commons.customselector.database.UploadedStatusDao
 import fr.free.nrw.commons.upload.depicts.Depicts
 import fr.free.nrw.commons.upload.depicts.DepictsDao
